requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
google-auth==2.6.0
idna==3.3
kubernetes==21.7.0
mysqlclient==2.1.0
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
websocket-client==1.2.3
django-cors-headers==3.10.0